reimplement inexact deduplication
My first attempt at inexact deduplication was plagued by a lot of false
negatives due to cases of duplicate which had perceptual hashes that were
similar but not exactly the same.

This new version handles such cases better by grouping together items with
similar p-hashes rather than just those with exactly the same p-hashes.
Moreover, we define "similarity" in such a way that if the p-hashes for two
images are not similar, then they are *guaranteed* not to be duplicates as
defined by the `deduplicate` function.

The new version is more computationally intensive, but also much more reliable.
Fortunately, it only runs in the background, and only when new items appear on
the filesystem, so users won't notice the performance difference in practice.

This commit is also an excuse to start using the new "captured identifiers in
format strings" feature introduced in Rust 1.58.0.