handle invalid or missing timestamps more gracefully
Previously, if we couldn't find a timestamp in the metadata embedded in a file
(or there was no metadata at all), we would add it to the `bad_paths` table with
no further consideration.  Now we first check to see if the file is part of a
Google Photos export, in which case the metadata may be found in a JSON file in
the same directory.  If that doesn't work, we use a default timestamp of
1970-01-01, which will cause the item to be tagged with "year:unknown".  We only
insert into `bad_paths` if the file can't be parsed and decoded.

Additionally, if we *can* find a timestamp in the metadata, we do a sanity check
to make sure e.g. it's not in the far future.  If it fails the sanity check, we
use the default timestamp instead.